Separate initial config flow
The json-file can be imported in nodered.
Open the function node 'set initial config' and edit according to the instructions in the file.
The goal of this flow is to initialize a new button plus device with some initials settings to enable it to be used with the flow in the main branch.
After editing the function, click the inject node. After a few seconds, a message will apear on you buttonplus and the internal temperature will be displayed. (In V0.80 that has an error, it does not work yet.)
The source of the two functions are also uploaded for reference. You do not need them.